refactor: insurance, proposal 폴더링 수정
